api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.135.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "sqlalchemy[asyncio]>=2.0.47",
]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
]
